Removed sections from HomePage to keep it simple.
